split citation concern to own controller
